change version and update some srcipts

+++ changethreads.sh
@@ -9,9 +9,16 @@ fi
 
 sudo DEBIAN_FRONTEND=noninteractive apt install jq -y
 
+qubic_folder="/home/qubic/qcli"
+if [[ -e $qubic_folder ]]; then
 jq --arg threads "$threads" \
     '.Settings.amountOfThreads = $threads' \
     /home/qubic/qcli/appsettings.json > tmp && mv tmp /home/qubic/qcli/appsettings.json
+else
+jq --arg threads "$threads" \
+    '.Settings.amountOfThreads = $threads' \
+    /root/qcli/appsettings.json > tmp && mv tmp /root/qcli/appsettings.json
+fi
 
 sudo systemctl restart qli.service
 sudo journalctl -u qli.service -f
